context_save: remove handle from handle_manager when saving sessions
Esys_ContextSave has a special handling for session handles where it
invalidates the matching ESYS_TR object as the TPM frees the
corresponding session slot after saving it.

As tss_esapi::Context still tries to flush the (now invalid) session
handle when dropped, this causes the following warnings to be emitted
by the TSS library:
> ERROR:esys:src/tss2-esys/esys_iutil.c:1096:esys_GetResourceObject()
> Error: Esys handle does not exist (70018).
> ERROR:esys:src/tss2-esys/api/Esys_FlushContext.c:138:Esys_FlushContext_Async()
> flushHandle unknown. ErrorCode (0x00070018)
> ERROR:esys:src/tss2-esys/api/Esys_FlushContext.c:66:Esys_FlushContext()
> Error in async function ErrorCode (0x00070018)

To prevent this, context_save now tries to identify whether the provided
object handle is a session handle, and if so drops the handle_manager
reference to this handle.

[1]
https://github.com/tpm2-software/tpm2-tss/blob/master/src/tss2-esys/api/Esys_ContextSave.c#L283

Signed-off-by: Julien Gomes <julien@arista.com>